requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-environ>=0.11,<1.0
django-cors-headers>=4.3,<5.0
django-ratelimit>=4.1,<5.0
python-json-logger>=2.0,<3.0

celery>=5.3,<6.0
redis>=5.0,<6.0

psycopg[binary]>=3.1,<4.0

PyMySQL>=1.1,<2.0

cryptography>=42.0,<46.0

pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
freezegun>=1.4,<2.0
factory-boy>=3.3,<4.0
